requirements.txt
Markdown==3.10.2
pandas==3.0.2
python-dotenv==1.2.2
duckdb==1.5.2
openpyxl==3.1.5
langchain==1.2.15
langchain-ollama==1.1.0
langchain-groq==1.1.2
langchain-google-genai==4.2.2
streamlit==1.57.0
fastapi==0.115.12
uvicorn==0.34.3
requests==2.32.3
langsmith
